pyproject.toml
[project]
name = "voice-input"
version = "0.1.0"
description = "本地语音输入助手（Ubuntu + vLLM ASR）"
requires-python = ">=3.10"
dependencies = [
    "python-xlib>=0.33",
    "sounddevice==0.5.1",
    "numpy==2.2.6",
    "requests==2.32.3",
    "pyperclip==1.9.0",
    "pyyaml==6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "requests-mock>=1.12",
    "scipy>=1.14",
]

[tool.setuptools]
py-modules = [
    "main",
    "state_machine",
    "audio_recorder",
    "asr_client",
    "text_injector",
    "hotkey_manager",
    "notifier",
    "utils",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
